+++ R/read_scan.R
@@ -83,18 +83,26 @@ read_parquet_polars <- function(
   hive_schema = NULL,
   try_parse_hive_dates = TRUE,
   glob = TRUE,
-  rechunk = TRUE,
   low_memory = FALSE,
   storage_options = NULL,
   use_statistics = TRUE,
   cache = TRUE,
-  include_file_paths = NULL
+  include_file_paths = NULL,
+  rechunk
 ) {
   rlang::arg_match0(
     parallel,
     values = c("auto", "columns", "row_groups", "prefiltered", "none")
   )
   rlang::check_dots_empty()
+
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "read_parquet_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
 
   scan_parquet_polars(
     source = source,
@@ -106,7 +114,6 @@ read_parquet_polars <- function(
     hive_schema = hive_schema,
     try_parse_hive_dates = try_parse_hive_dates,
     glob = glob,
-    rechunk = rechunk,
     low_memory = low_memory,
     storage_options = storage_options,
     use_statistics = use_statistics,
@@ -130,18 +137,26 @@ scan_parquet_polars <- function(
   hive_schema = NULL,
   try_parse_hive_dates = TRUE,
   glob = TRUE,
-  rechunk = FALSE,
   low_memory = FALSE,
   storage_options = NULL,
   use_statistics = TRUE,
   cache = TRUE,
-  include_file_paths = NULL
+  include_file_paths = NULL,
+  rechunk
 ) {
   rlang::arg_match0(
     parallel,
     values = c("auto", "columns", "row_groups", "prefiltered", "none")
   )
   rlang::check_dots_empty()
+
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "scan_parquet_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
 
   pl$scan_parquet(
     source = source,
@@ -153,7 +168,6 @@ scan_parquet_polars <- function(
     hive_schema = hive_schema,
     try_parse_hive_dates = try_parse_hive_dates,
     glob = glob,
-    rechunk = rechunk,
     low_memory = low_memory,
     storage_options = storage_options,
     use_statistics = use_statistics,
@@ -170,10 +184,6 @@ scan_parquet_polars <- function(
 #' `scan_csv_polars()` imports the data as a Polars LazyFrame.
 #'
 #' @inherit polars::pl__scan_csv params details
-#' @param dtypes `r lifecycle::badge("deprecated")` Deprecated,
-#' use `schema_overrides` instead.
-#' @param reuse_downloaded `r lifecycle::badge("deprecated")`
-#' Deprecated with no replacement.
 #'
 #' @rdname from_csv
 #' @name from_csv
@@ -266,7 +276,6 @@ read_csv_polars <- function(
   n_rows = NULL,
   encoding = "utf8",
   low_memory = FALSE,
-  rechunk = TRUE,
   skip_rows_after_header = 0,
   row_index_name = NULL,
   row_index_offset = 0,
@@ -275,24 +284,22 @@ read_csv_polars <- function(
   raise_if_empty = TRUE,
   truncate_ragged_lines = FALSE,
   include_file_paths = NULL,
-  dtypes,
-  reuse_downloaded
+  rechunk
 ) {
   rlang::arg_match0(encoding, values = c("utf8", "utf8-lossy"))
   rlang::check_dots_empty()
 
-  if (!missing(dtypes)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "read_csv_polars(dtypes)",
-      details = "Use `schema_overrides` instead."
+      when = "0.20.0",
+      what = "read_csv_polars(rechunk)",
+      details = "This argument has no replacement."
     )
-    schema_overrides <- dtypes
   }
-  if (!missing(reuse_downloaded)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "read_csv_polars(reuse_downloaded)",
+      when = "0.20.0",
+      what = "read_csv_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -313,7 +320,6 @@ read_csv_polars <- function(
     n_rows = n_rows,
     encoding = encoding,
     low_memory = low_memory,
-    rechunk = rechunk,
     skip_rows_after_header = skip_rows_after_header,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
@@ -346,7 +352,6 @@ scan_csv_polars <- function(
   n_rows = NULL,
   encoding = "utf8",
   low_memory = FALSE,
-  rechunk = TRUE,
   skip_rows_after_header = 0,
   row_index_name = NULL,
   row_index_offset = 0,
@@ -355,24 +360,22 @@ scan_csv_polars <- function(
   raise_if_empty = TRUE,
   truncate_ragged_lines = FALSE,
   include_file_paths = NULL,
-  dtypes,
-  reuse_downloaded
+  rechunk
 ) {
   rlang::arg_match0(encoding, values = c("utf8", "utf8-lossy"))
   rlang::check_dots_empty()
 
-  if (!missing(dtypes)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "scan_csv_polars(dtypes)",
-      details = "Use `schema_overrides` instead."
+      when = "0.20.0",
+      what = "scan_csv_polars(rechunk)",
+      details = "This argument has no replacement."
     )
-    schema_overrides <- dtypes
   }
-  if (!missing(reuse_downloaded)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "scan_csv_polars(reuse_downloaded)",
+      when = "0.20.0",
+      what = "scan_csv_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -393,7 +396,6 @@ scan_csv_polars <- function(
     n_rows = n_rows,
     encoding = encoding,
     low_memory = low_memory,
-    rechunk = rechunk,
     skip_rows_after_header = skip_rows_after_header,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
@@ -413,9 +415,6 @@ scan_csv_polars <- function(
 #' `scan_ndjson_polars()` imports the data as a Polars LazyFrame.
 #'
 #' @inherit polars::pl__scan_ndjson params details
-#' @param reuse_downloaded `r lifecycle::badge("deprecated")`
-#' Deprecated with no replacement.
-#'
 #'
 #' @rdname from_ndjson
 #' @name from_ndjson
@@ -483,17 +482,23 @@ read_ndjson_polars <- function(
   batch_size = NULL,
   n_rows = NULL,
   low_memory = FALSE,
-  rechunk = FALSE,
   row_index_name = NULL,
   row_index_offset = 0,
   ignore_errors = FALSE,
-  reuse_downloaded
+  rechunk
 ) {
   rlang::check_dots_empty()
-  if (!missing(reuse_downloaded)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "read_ndjson_polars(reuse_downloaded)",
+      when = "0.20.0",
+      what = "read_ndjson_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "read_ndjson_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -504,7 +509,6 @@ read_ndjson_polars <- function(
     batch_size = batch_size,
     n_rows = n_rows,
     low_memory = low_memory,
-    rechunk = rechunk,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
     ignore_errors = ignore_errors
@@ -522,17 +526,23 @@ scan_ndjson_polars <- function(
   batch_size = NULL,
   n_rows = NULL,
   low_memory = FALSE,
-  rechunk = FALSE,
   row_index_name = NULL,
   row_index_offset = 0,
   ignore_errors = FALSE,
-  reuse_downloaded
+  rechunk
 ) {
   rlang::check_dots_empty()
-  if (!missing(reuse_downloaded)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "scan_ndjson_polars(reuse_downloaded)",
+      when = "0.20.0",
+      what = "scan_ndjson_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "scan_ndjson_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -543,7 +553,6 @@ scan_ndjson_polars <- function(
     batch_size = batch_size,
     n_rows = n_rows,
     low_memory = low_memory,
-    rechunk = rechunk,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
     ignore_errors = ignore_errors
@@ -558,8 +567,6 @@ scan_ndjson_polars <- function(
 #' `scan_ipc_polars()` imports the data as a Polars LazyFrame.
 #'
 #' @inherit polars::pl__scan_ipc params details
-#' @param memory_map `r lifecycle::badge("deprecated")` Deprecated
-#' with no replacement.
 #'
 #' @rdname from_ipc
 #' @name from_ipc
@@ -626,16 +633,22 @@ read_ipc_polars <- function(
   n_rows = NULL,
   row_index_name = NULL,
   row_index_offset = 0L,
-  rechunk = FALSE,
   cache = TRUE,
   include_file_paths = NULL,
-  memory_map
+  rechunk
 ) {
   rlang::check_dots_empty()
-  if (!missing(memory_map)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "read_ipc_polars(memory_map)",
+      when = "0.20.0",
+      what = "read_ipc_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "read_ipc_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -645,7 +658,6 @@ read_ipc_polars <- function(
     n_rows = n_rows,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
-    rechunk = rechunk,
     cache = cache,
     include_file_paths = include_file_paths
   ) |>
@@ -661,16 +673,22 @@ scan_ipc_polars <- function(
   n_rows = NULL,
   row_index_name = NULL,
   row_index_offset = 0L,
-  rechunk = FALSE,
   cache = TRUE,
   include_file_paths = NULL,
-  memory_map
+  rechunk
 ) {
   rlang::check_dots_empty()
-  if (!missing(memory_map)) {
+  if (!missing(rechunk)) {
     lifecycle::deprecate_warn(
-      when = "0.14.0",
-      what = "scan_ipc_polars(memory_map)",
+      when = "0.20.0",
+      what = "scan_ipc_polars(rechunk)",
+      details = "This argument has no replacement."
+    )
+  }
+  if (!missing(rechunk)) {
+    lifecycle::deprecate_warn(
+      when = "0.20.0",
+      what = "scan_ipc_polars(rechunk)",
       details = "This argument has no replacement."
     )
   }
@@ -680,7 +698,6 @@ scan_ipc_polars <- function(
     n_rows = n_rows,
     row_index_name = row_index_name,
     row_index_offset = row_index_offset,
-    rechunk = rechunk,
     cache = cache,
     include_file_paths = include_file_paths
   )