+++ tests/testthat/test-read-write.R
@@ -85,20 +85,11 @@ test_that("deprecated arguments in scan/read_csv_polars", {
   x <- write_csv_polars(dat_pl, dest)
 
   expect_snapshot({
-    new_dat <- scan_csv_polars(dest, dtypes = list(drat = pl$Float32))
+    x <- scan_csv_polars(dest, rechunk = TRUE)
   })
   expect_snapshot({
-    x <- scan_csv_polars(dest, reuse_downloaded = TRUE)
+    x <- read_csv_polars(dest, rechunk = TRUE)
   })
-  expect_true(new_dat$collect_schema()[["drat"]]$eq(pl$Float32))
-
-  expect_snapshot({
-    new_dat <- read_csv_polars(dest, dtypes = list(drat = pl$Float32))
-  })
-  expect_snapshot({
-    x <- read_csv_polars(dest, reuse_downloaded = TRUE)
-  })
-  expect_true(new_dat$collect_schema()[["drat"]]$eq(pl$Float32))
 })
 
 test_that("deprecated arguments in scan/read_ipc_polars", {
@@ -107,10 +98,10 @@ test_that("deprecated arguments in scan/read_ipc_polars", {
   x <- write_ipc_polars(dat_pl, dest)
 
   expect_snapshot({
-    new_dat <- scan_ipc_polars(dest, memory_map = TRUE)
+    new_dat <- scan_ipc_polars(dest, rechunk = TRUE)
   })
   expect_snapshot({
-    new_dat <- read_ipc_polars(dest, memory_map = TRUE)
+    new_dat <- read_ipc_polars(dest, rechunk = TRUE)
   })
 })
 
@@ -120,9 +111,22 @@ test_that("deprecated arguments in scan/read_ndjson_polars", {
   x <- write_ndjson_polars(dat_pl, dest)
 
   expect_snapshot({
-    new_dat <- scan_ndjson_polars(dest, reuse_downloaded = TRUE)
+    new_dat <- scan_ndjson_polars(dest, rechunk = TRUE)
   })
   expect_snapshot({
-    new_dat <- read_ndjson_polars(dest, reuse_downloaded = TRUE)
+    new_dat <- read_ndjson_polars(dest, rechunk = TRUE)
+  })
+})
+
+test_that("deprecated arguments in scan/read_parquet_polars", {
+  dat_pl <- as_polars_df(mtcars)
+  dest <- tempfile(fileext = ".parquet")
+  x <- write_parquet_polars(dat_pl, dest)
+
+  expect_snapshot({
+    new_dat <- scan_parquet_polars(dest, rechunk = TRUE)
+  })
+  expect_snapshot({
+    new_dat <- read_parquet_polars(dest, rechunk = TRUE)
   })
 })